+++ R/zzz.R
@@ -1,0 +1,17 @@
+
+utils::globalVariables(
+  c('f_icon',
+    'n_valid',
+    'perc_valid',
+    'n_unique',
+    'perc_unique',
+    'n_zero',
+    'perc_zero',
+    'n_nas',
+    'perc_nas',
+    'size',
+    'preview',
+    'q1',
+    'q3'
+    )
+  )